+++ list.h
@@ -6,12 +6,12 @@ class List {
         virtual ~List() {};
 
         //virtual int &get(char symbol) const = 0;
-        virtual const int &get(char symbol) const = 0;
-        virtual int &get(char symbol) = 0;
-        virtual void set(char symbol, int value) = 0;
+        virtual const float &get(char symbol) const = 0;
+        virtual float &get(char symbol) = 0;
+        virtual void set(char symbol, float value) = 0;
 
-        virtual void insert(int i, char symbol, int value) = 0;
-        virtual void append(char symbol, int value) = 0;
+        virtual void insert(int i, char symbol, float value) = 0;
+        virtual void append(char symbol, float value) = 0;
         virtual void remove(char symbol) = 0;
         virtual void clear() = 0;
